@@ -1,7 +1,7 @@
 package br.com.developer.allefsousa.adorofilmes.utils;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 
